rename and move datasetBuilderUtils and test
