+++ src/dataset-builder/DomainCriteriaSelector.ts
@@ -1,6 +1,7 @@
 import _ from 'lodash/fp';
 import { h } from 'react-hyperscript-helpers';
 import { spinnerOverlay } from 'src/components/common';
+import { DomainCriteria, GetConceptsResponse } from 'src/dataset-builder/DatasetBuilderUtils';
 import {
   DataRepo,
   SnapshotBuilderConcept as Concept,
@@ -8,7 +9,6 @@ import {
 } from 'src/libs/ajax/DataRepo';
 import { useLoadedData } from 'src/libs/ajax/loaded-data/useLoadedData';
 import { useOnMount } from 'src/libs/react-utils';
-import { DomainCriteria, GetConceptsResponse } from 'src/libs/util/DatasetBuilder';
 
 import { ConceptSelector } from './ConceptSelector';
 import { cohortEditorState, DomainCriteriaSelectorState } from './dataset-builder-types';
