+++ src/dataset-builder/dataset-builder-types.ts
@@ -1,5 +1,5 @@
+import { Cohort, CriteriaGroup } from 'src/dataset-builder/DatasetBuilderUtils';
 import { SnapshotBuilderDomainOption as DomainOption } from 'src/libs/ajax/DataRepo';
-import { Cohort, CriteriaGroup } from 'src/libs/util/DatasetBuilder';
 
 let groupCount = 1;
 export const newCriteriaGroup = (): CriteriaGroup => {
